move extract to utils
